Renamed Documentator by Documentor
Renamed all occurences of ```Documentator``` by ```Documentor``` to avoid break compatibility.
